Numerous Ambisonic bug fixes
Moves the handling of the Ambisonic mode switching from soundObject.js to main.js: this allows the browser output channel count check to be deffered, which previously caused an error with Safari. Furthermore, this enables sound elements to be initialized for binaural or ambisonic based on current output mode. Also allows zones to be merged to W component of B-format signal in ambisonic mode to be heard omnidirectionally (previously zones remained binaural, which targeted the first two channels in ambisonic mode, resulting in left-panned zones). Sphere plus cone sounds and zone sounds are now collated in different arrays to be passed to main and processed accordingly based on output mode. Numerous referencing changes to support this implementation. Also changes the way output mode help text is displayed.